refactor : `exception` 패키지를 `common` 하위로 이동
