+++ src/main/java/kitchenpos/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package kitchenpos;
 
-import kitchenpos.exception.CannotFindException;
-import kitchenpos.exception.CannotUpdateException;
+import kitchenpos.common.exception.CannotFindException;
+import kitchenpos.common.exception.CannotUpdateException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
