+++ src/main/java/kitchenpos/common/exception/CannotFindException.java
@@ -1,6 +1,6 @@
-package kitchenpos.exception;
+package kitchenpos.common.exception;
 
-import kitchenpos.common.Message;
+import kitchenpos.common.exception.Message;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
