+++ src/main/java/kitchenpos/ordertable/application/TableService.java
@@ -1,6 +1,6 @@
 package kitchenpos.ordertable.application;
 
-import kitchenpos.exception.CannotFindException;
+import kitchenpos.common.exception.CannotFindException;
 import kitchenpos.ordertable.event.TableEmptyStatusChangedEvent;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.domain.OrderTable;
@@ -13,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static kitchenpos.common.Message.ERROR_ORDER_TABLE_NOT_FOUND;
+import static kitchenpos.common.exception.Message.ERROR_ORDER_TABLE_NOT_FOUND;
 
 @Service
 @Transactional
